Upgrade to OpenJDK 7 required by CommonAPI
Upgrading to JDK 7 is long overdue - we only used JDK6 because a
much earlier Franca version suggested it.

This was already done on master a long time ago anyhow but a merge is
long overdue between cpp_commonapi branch and master branches.

CommonAPI C++ tools (at least the now used version) requires JDK7. It
seemed to fail silently on JDK6 so that the context menus in Eclipse do not
even appear. The menus now appear and code generation is possible.

Not sure about JDK8 compatibility yet - interested if someone can test.