Remove installation of PTP eclipse.remote again
This reverts commit 12c61b2b5784cbc87f2d5739a247c194cf51d585.
"Change order of installation PTP/CDT"

The commit message was basically wrong - order was not changed but
org.eclipse.remote... was added back.  However that prerequisite is not
there at the moment - maybe later when more CDT / Linuxtools added back.

Amended also removal of some remnants of it inside CONFIG