Do java version check
This is mostly useful when running script.sh directly on host since
Vagrantfile ought to ensure it is the right one for any given branch,
unless we make a mistake.  For the direct host install it is however
necessary (but not always sufficient to guarantee it works anyway!).

A MD5SUM variable change accidentally came along with the cherry pick...