Generalize install() function to auto handle different types
Generalize install() function to look at the CONFIG variables to figure out
if this is an update_site type of installation or an archive file.
Then call the appropriate function.  Now, to change installation type
just change the config and no need to also change the script (except for
the special cases that remain).