Force virtualbox as default provider in Vagrantfile
In a system where other providers (e.g. libvirt)
are available, it seems that Vagrant will default to those.
I had never noticed it before but ran into this problem now
where libvirt was chosen instead, and some errors appeared.

With this ENV variable set directly in the file we force the Vagrant
provider to be VirtualBox which is the only tested configuration.