Update Vagrant base box to Ubuntu 16.04
This supports Franca 11 <-> Eclipse Neon <-> Java 8 combination. Since
Java 8 is the default now in 16.04 it removes the need to use a special
PPA for openjdk.

N.B. This is an experimental branch.  The Franca packages are not
guaranteed to work with Neon apparently (even if they seem to
install OK)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>